odb-export: initial version
For now just dump a named object to stdout. The goal is to have a
pair of services for import/export that can exchange information
about shared objects to build a minimal set of data to transfer,
akin to git's pack protocol. A degenerate case of this is when
there are no shared objects and the entirety of the database is
transferred, which could be useful for schema changes.