odb-insert: initial support for extracting tarball
Adds some initial support for representing directory trees in the
database. This does not support all of the metadata present in the
header (times, modes, ownership, etc) but is enough to populate the
database with a number of objects and be able to see the links between
them. For example:
  $ HASH=fad8c57097ee1c8325e1b4e9ffe8a1dafad7646abd214ab62a2a4909439b93cf
  $ cargo run --bin odb-export -- --db my.db $HASH | \
        capnp convert binary:json proto/odb.capnp Object