capnp: add LazyTypedReader utility
Continuing to work around issues found when using capnp in Sync
contexts. Messages are not able to be shared across threads because of
the stateful nature of consuming them (pointer bounds validation,
mostly). There exists a sync version of the reader via the sync_reader
feature, and it might be worth playing with that in the future. In the
interim, this is a version that produces the !Sync reader on-demand from
Sync backing storage, which handles at least the case of objects read
from the database.