fuse: plumb through stat structures
This is almost sufficient to allow the first stat() call on the root
of the filesystem to work end-to-end. This is because the schema has the
stat data living on the directory entry and not the directory itself.
This is what we want in most cases, but not at the root. I haven't
decided how I want to handle this yet, so right now the root stat is
just hardcoded. We should be able to see more interesting data once we
hook up readdir.